requirements.txt
fastapi
uvicorn
playwright
python-multipart
asyncio
openai
python-dotenv
tiktoken
